S02E02_analyze_images/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0